RoR2 Logic (With Messy Debug)
Allow RoR2 to create logic for unlocking environments. This logic moves
all the environments into their orderedstage order which can then be
used to see if full loop can be complted and then other endings. All
environments are considered progression items. Also grant a full loop of
stages to the player otherwise players may have to resetart runs to go
back and unlock necessary environments.
(Also add tons and tons of debug which needs to be cleaned up and/or
removed.)